requirements.txt
pyrealsense2
numpy
opencv-python
pupil-apriltags
scipy
pyyaml
